+++ src/settings/MappingProfiles/MappingProfilesContainer/MappingProfilesContainer.js
@@ -14,7 +14,7 @@ import {
   location as locationShape,
 } from 'react-router-prop-types';
 
-import { Route, TitleManager } from '@folio/stripes/core';
+import { Route, TitleManager, useStripes } from '@folio/stripes/core';
 import { makeQueryFunction } from '@folio/stripes/smart-components';
 import {
   MappingProfiles,
@@ -69,6 +69,7 @@ const MappingProfilesContainer = ({
   mutator,
 }) => {
   const intl = useIntl();
+  const stripes = useStripes();
   const allTransformations = useMemo(
     () => orderBy(get(resources.transformations.records, '0.transformationFields', [])
       .map(transformation => ({
@@ -81,6 +82,8 @@ const MappingProfilesContainer = ({
     [intl, resources.transformations.records]
   );
   const isTransformationsLoaded = get(resources, 'transformations.hasLoaded', false);
+  const hasViewPerms = stripes.hasPerm('settings.data-export.view') && !stripes.hasPerm('module.data-export.enabled');
+  const lastMenu = hasViewPerms ? (<></>) : '';
 
   const handleNavigationToMappingProfilesList = useCallback(
     () => push(`${path}${search}`),
@@ -101,6 +104,7 @@ const MappingProfilesContainer = ({
         <MappingProfiles
           parentResources={resources}
           parentMutator={mutator}
+          lastMenu={lastMenu}
           formatter={useMappingProfileListFormatter({ format: ({ outputFormat }) => outputFormat })}
           {...useMappingProfilesProperties(customProperties)}
         />